requirements.txt
altair==4.2.0
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
catboost==1.0.6
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
Django==4.0.6
entrypoints==0.4
executing==0.10.0
fonttools==4.36.0
gitdb==4.0.9
GitPython==3.1.27
graphviz==0.20.1
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
imbalanced-learn==0.9.1
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.14.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-widgets==3.0.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
missingno==0.5.1
multimethod==1.8
nest-asyncio==1.5.5
networkx==2.8.5
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pandas-profiling==3.2.0
parso==0.8.3
phik==0.12.2
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyarrow==9.0.0
pydantic==1.9.2
pydeck==0.7.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywedge==0.5.1.8
pywin32==304
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
rich==12.5.1
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
stack-data==0.4.0
streamlit==1.12.0
streamlit-pandas-profiling==0.1.3
tangled-up-in-unicode==0.2.0
tenacity==8.0.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
visions==0.7.4
watchdog==2.1.9
wcwidth==0.2.5
widgetsnbextension==4.0.2
xgboost==1.6.1
zipp==3.8.1
